add main of func and all

+++ libft/ft_itoa.c
@@ -6,13 +6,13 @@
 /*   By: achepurn <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/04 12:50:08 by achepurn          #+#    #+#             */
-/*   Updated: 2017/11/20 18:20:47 by achepurn         ###   ########.fr       */
+/*   Updated: 2017/12/18 15:14:15 by achepurn         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-static size_t	a_len(int n)
+static size_t	a_len(intmax_t n)
 {
 	size_t	res;
 
@@ -29,9 +29,9 @@ static size_t	a_len(int n)
 	return (res);
 }
 
-static void		put_a(int n, char *str, int len)
+static void		put_a(intmax_t n, char *str, size_t len)
 {
-	unsigned int	nb;
+	uintmax_t	nb;
 
 	if (n < 0)
 		str[0] = '-';
@@ -47,7 +47,7 @@ static void		put_a(int n, char *str, int len)
 	str[len] = nb + '0';
 }
 
-char			*ft_itoa(int n)
+char			*ft_itoa(intmax_t n)
 {
 	char	*res;
 	size_t	len;
